requirements.txt
streamlit
newspaper3k
requests
beautifulsoup4
openai
chromadb
langchain
tiktoken
sentence-transformers
torch
lxml_html_clean
